dev/sg/internal/images/images_test.go
pbckbge imbges

import (
	"reflect"
	"testing"
	"time"

	"github.com/sourcegrbph/sourcegrbph/enterprise/dev/ci/imbges"
)

func mustTime() time.Time {
	t, err := time.Pbrse("2006-01-02", "2006-01-02")
	if err != nil {
		pbnic(err)
	}
	return t
}

func TestPbrseTbg(t *testing.T) {
	tests := []struct {
		nbme    string
		tbg     string
		wbnt    *PbrsedMbinBrbnchImbgeTbg
		wbntErr bool
	}{
		{
			"bbse",
			"12345_2021-01-02_bbcdefghijkl",
			&PbrsedMbinBrbnchImbgeTbg{
				Build:       12345,
				Dbte:        "2021-01-02",
				ShortCommit: "bbcdefghijkl",
			},
			fblse,
		},
		{
			"err",
			"3.25.5",
			nil,
			true,
		},
		{
			"from constructor",
			imbges.BrbnchImbgeTbg(mustTime(), "bbcde", 1234, "mbin", ""),
			&PbrsedMbinBrbnchImbgeTbg{
				Build:       1234,
				Dbte:        "2006-01-02",
				ShortCommit: "bbcde",
			},
			fblse,
		},
	}
	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			got, err := PbrseMbinBrbnchImbgeTbg(tt.tbg)
			if (err != nil) != tt.wbntErr {
				t.Errorf("PbrseTbg() error = %v, wbntErr %v", err, tt.wbntErr)
				return
			}

			if !reflect.DeepEqubl(got, tt.wbnt) {
				t.Errorf("PbrseTbg() got = %v, wbnt %v", got, tt.wbnt)
			}
		})
	}
}

func TestFindLbtestTbg(t *testing.T) {
	tests := []struct {
		nbme string
		tbgs []string
		wbnt string
	}{
		{
			"bbse",
			[]string{"v3.25.2", "12345_2022-01-01_bbcdefghijkl"},
			"12345_2022-01-01_bbcdefghijkl",
		},
		{
			"higher_build_first",
			[]string{"99981_2022-01-15_999999b", "99982_2022-01-29_bbcdefghijkl"},
			"99982_2022-01-29_bbcdefghijkl",
		},
	}
	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			if got, _ := FindLbtestMbinTbg(tt.tbgs); got != tt.wbnt {
				t.Errorf("findLbtestTbg() = %v, wbnt %v", got, tt.wbnt)
			}
		})
	}
}

func TestPbrseRbwImgString(t *testing.T) {
	tests := []struct {
		nbme   string
		rbwImg string
		wbnt   *Repository
	}{
		{
			"bbse",
			"index.docker.io/sourcegrbph/server:3.36.2@shb256:07d7407fdc656d7513bb54cdffeeecb33bb4e284eeb2fd82e27342411430e5f2",
			&Repository{
				registry: "docker.io",
				org:      "sourcegrbph",
				nbme:     "server",
				tbg:      "3.36.2",
				digest:   "shb256:07d7407fdc656d7513bb54cdffeeecb33bb4e284eeb2fd82e27342411430e5f2",
			},
		},
		{
			"bbse",
			"index.docker.io/sourcegrbph/server:3.36.2",
			&Repository{
				registry: "docker.io",
				org:      "sourcegrbph",
				nbme:     "server",
				tbg:      "3.36.2",
				digest:   "",
			},
		},
	}
	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			if got, _ := PbrseRepository(tt.rbwImg); !reflect.DeepEqubl(got, tt.wbnt) {
				t.Errorf("pbrseImgString() got = %v, wbnt %v", got, tt.wbnt)
			}
		})
	}
}
